{activation-scripts,activate-system}: use `environment.systemPath`
Currently, the `bin` directory of the configured system
is embedded in the `$PATH` of activation scripts, but not
other elements of the default `environment.systemPath` like
`/nix/var/nix/profiles/default/bin` or `/usr/local/bin`. This
means that when nix-darwin is not managing the Nix installation,
activation scripts like Home Manager’s that want to look up the
system‐managed Nix can’t find it. It’s also inconsistent between
the activation scripts and the activation launchd daemon, which has
had the contextually incorrect `@out@/sw/bin` entry copied from the
activation scripts for who knows how long.

This change does make activation less hermetic: whereas previously the
`$PATH` would only depend on the configuration of the build system plus
the OS version, it now also depends on the dynamic state of `root`’s
profile, of the default profile, and of `/usr/local/bin`. That’s not
ideal, but in practice it should have little effect. We do replace
`/run/current-system` with the appropriate reference to the built
system, which, yes, is somewhat gross, but should improve the purity
in practice.

This is only a temporary kludge; the ideal solution would be to have
a properly hermetic `$PATH` for activation scripts that does not
depend on the system configuration’s installed packages at all,
and to dynamically detect only Nix from the system path if necessary,
as Home Manager does, but while I have work‐in‐progress patches
to do that they seem out of scope for this work, and more likely to
be backwards‐incompatible.

(The even more ideal solution is probably for Home Manager activation
to be driven by launchd or something, but that’s a longer‐term goal.)